Brody code mix

+++ Core/Inc/distance.h
@@ -30,7 +30,7 @@ void DistanceInit(ADC_HandleTypeDef *_adc, DMA_HandleTypeDef *_dma){
 	adc = *_adc;
 	dma = *_dma;
 
-	HAL_ADC_Start_DMA(&adc,(uint32_t*)&rawADCValues,3);
+	HAL_ADC_Start_DMA(&adc,(uint16_t*)&rawADCValues,3);
 	ADC1->CR2 |= ADC_CR2_DDS;
 }
